refactor: rename fazdb to fazbeat/api_collect
Rename fazdb to fazbeat/api_collect to avoid confusion with the database
"faz-db". Future heartbeat packages will also be developed here as a
subdirectory of fazbeat.
