+++ .env-example
@@ -5,8 +5,8 @@ DEV_SERVER_ID=0
 
 FAZCORD_DISCORD_LOG_WEBHOOK=""
 FAZCORD_DISCORD_STATUS_WEBHOOK=""
-FAZDB_DISCORD_LOG_WEBHOOK=""
-FAZDB_DISCORD_STATUS_WEBHOOK=""
+API_COLLECTOR_DISCORD_LOG_WEBHOOK=""
+API_COLLECTOR_DISCORD_STATUS_WEBHOOK=""
 
 FAZCORD_DB_MAX_RETRIES=3
-FAZDB_DB_MAX_RETRIES=3
+API_COLLECTOR_DB_MAX_RETRIES=3
