+++ .github/workflows/release.yml
@@ -96,8 +96,8 @@ jobs:
             commit=${{ github.sha }}
             build_date=${{ env.date }}
 
-  build-fazdb:
-    name: Build fazdb Docker Images
+  build-api_collect:
+    name: Build api_collect Docker Images
     needs: [changelog] # Build needs the new version number
     runs-on: ubuntu-latest
     steps:
@@ -126,7 +126,7 @@ jobs:
         with:
           name: version
 
-      - name: Build and Push fazdb
+      - name: Build and Push api_collect
         uses: docker/build-push-action@v6
         with:
           context: .
@@ -134,8 +134,8 @@ jobs:
           platforms: linux/amd64
           push: true
           tags: |
-            ghcr.io/fazuh/fazdb:latest
-            ghcr.io/fazuh/fazdb:${{ needs.changelog.outputs.tag }}
+            ghcr.io/fazuh/api_collect:latest
+            ghcr.io/fazuh/api_collect:${{ needs.changelog.outputs.tag }}
           labels: |
             version=${{ needs.changelog.outputs.tag }}
             commit=${{ github.sha }}
@@ -143,7 +143,7 @@ jobs:
 
   release-github:
     name: Release to GitHub
-    needs: [build-fazcord, build-fazdb, changelog]
+    needs: [build-fazcord, build-api_collect, changelog]
     if: ${{ needs.changelog.outputs.skipped != 'true' }}
     runs-on: ubuntu-latest
     steps:
@@ -180,7 +180,7 @@ jobs:
 
   deploy:
     name: Deploy to Server
-    needs: [build-fazcord, build-fazdb, changelog]
+    needs: [build-fazcord, build-api_collect, changelog]
     runs-on: ubuntu-latest
     if: ${{ needs.changelog.outputs.skipped != 'true' }}
     steps:
