+++ docker/Dockerfile.api_collect
@@ -1,0 +1,11 @@
+FROM python:3.12.4-slim
+
+WORKDIR /app
+
+COPY fazbeat/api_collect ./fazbeat/api_collect
+COPY fazutil ./fazutil
+COPY requirements-api_collect.txt requirements.txt
+
+RUN pip3 install --no-cache-dir -r requirements.txt
+
+CMD ["python", "-m", "fazbeat.api_collect"]
