--- docker/Dockerfile.db
@@ -1,11 +1,0 @@
-FROM python:3.12.4-slim
-
-WORKDIR /app
-
-COPY fazdb ./fazdb
-COPY fazutil ./fazutil
-COPY requirements_fazdb.txt requirements.txt
-
-RUN pip3 install --no-cache-dir -r requirements.txt
-
-CMD ["python", "-m", "fazdb"]
