+++ fazbeat/api_collect/app/properties.py
