+++ fazbeat/api_collect/heartbeat/task/api_response_adapter.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime
-from typing import TYPE_CHECKING, Generator
+from typing import TYPE_CHECKING
 
 from fazutil.db.fazdb.model.character_history import CharacterHistory
 from fazutil.db.fazdb.model.character_info import CharacterInfo
