+++ requirements-fazcord.txt
