+++ tests/fazbeat/api_collector/test_api_response_adapter_fromfixture.py
