+++ tests/fazbeat/api_collector/test_request_queue.py
@@ -1,9 +1,8 @@
-from collections.abc import Coroutine
 from datetime import datetime as dt
 from unittest import TestCase
 from unittest.mock import MagicMock
 
-from fazdb.heartbeat.task.request_queue import RequestQueue
+from fazbeat.api_collect.heartbeat.task.request_queue import RequestQueue
 
 
 class TestRequestList(TestCase):
