+++ tests/fazbeat/api_collector/test_response_handler.py
@@ -2,8 +2,8 @@ from datetime import datetime
 from unittest import TestCase
 from unittest.mock import MagicMock
 
-from fazdb.heartbeat.task._response_handler import ResponseHandler
-from fazdb.heartbeat.task.request_queue import RequestQueue
+from fazbeat.api_collect.heartbeat.task._response_handler import ResponseHandler
+from fazbeat.api_collect.heartbeat.task.request_queue import RequestQueue
 from fazutil.api.wynn.wynn_api import WynnApi
