+++ tests/fazbeat/api_collector/test_response_queue.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import MagicMock
 
-from fazdb.heartbeat.task.response_queue import ResponseQueue
+from fazbeat.api_collect.heartbeat.task.response_queue import ResponseQueue
 
 
 class TestResponseList(unittest.TestCase):
