--- tests/fazdb/__main__.py
@@ -1,11 +1,0 @@
-# import unittest
-# import sys
-#
-# from loguru import logger
-#
-#
-# if __name__ == "__main__":
-#     suite = unittest.TestLoader().loadTestsFromTestCase(TestWorldsRepository)
-#     logger.debug(suite.countTestCases())
-#     runner = unittest.TextTestRunner(stream=sys.stderr, verbosity=2)
-#     runner.run(suite)
